refactoring of timon resources
